requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
Flask-Limiter==3.5.0
requests==2.31.0
python-dotenv==1.0.0
Werkzeug==3.0.1
Jinja2==3.1.5
MarkupSafe==3.0.2
itsdangerous==2.2.0
